포인트적립/탄소절약

+++ creen_carry_back/src/main/java/kr/co/iei/member/model/dao/MemberDao.java
@@ -44,7 +44,7 @@ public interface MemberDao {
 
 	void deleteMember(String memberId);
 
-	int getTotalCarbonPoint(String memberId);
+	Double getTotalCarbonPoint(String memberId);
 
 	int getCommunityTotalCarbon();
