+++ creen_carry_back/src/main/java/kr/co/iei/member/model/service/MemberService.java
@@ -223,7 +223,7 @@ public class MemberService {
         memberDao.deleteMember(memberId);
     }
 
-    public int getTotalCarbonPoint(String memberId) {
+    public Double getTotalCarbonPoint(String memberId) {
         return memberDao.getTotalCarbonPoint(memberId);
     }
